Added Dataless Annotators to the pipeline + some redundant code removal
