+++ dataless-classifier/pom.xml
@@ -3,7 +3,7 @@
     <parent>
         <artifactId>illinois-cogcomp-nlp</artifactId>
         <groupId>edu.illinois.cs.cogcomp</groupId>
-        <version>3.1.31</version>
+        <version>3.1.32</version>
     </parent>
 
     <modelVersion>4.0.0</modelVersion>
@@ -16,12 +16,12 @@
     	<dependency>
       		<groupId>edu.illinois.cs.cogcomp</groupId>
       		<artifactId>illinois-core-utilities</artifactId>
-      		<version>3.1.31</version>
+      		<version>3.1.32</version>
       	</dependency>
       	<dependency>
       		<groupId>edu.illinois.cs.cogcomp</groupId>
       		<artifactId>illinois-tokenizer</artifactId>
-      		<version>3.1.31</version>
+      		<version>3.1.32</version>
       	</dependency>
       	<dependency>
       		<groupId>log4j</groupId>
