+++ dataless-classifier/src/main/java/edu/illinois/cs/cogcomp/datalessclassification/util/QueryPreProcessor.java
@@ -17,22 +17,11 @@ public class QueryPreProcessor {
 	public static String process (String query) {
 		
 		if (stopSet.size() == 0) {
-			stopSet = Utilities.getStopWords();
+			stopSet = DatalessUtilities.getStopWords();
 		}
 		
 		String newQuery  = "";
 		
-//	    "(" ...
-//	    "*" ...
-//	    <QUOTED> ...
-//	    <TERM> ...
-//	    <PREFIXTERM> ...
-//	    <WILDTERM> ...
-//	    "[" ...
-//	    "{" ...
-//	    <NUMBER> ...
-//	    <TERM> ...
-//	    "*" ...
 		query = query.toLowerCase().replaceAll(",", " ").replaceAll(":", " ").replaceAll("\\.", " ");
 		query = query.toLowerCase().replaceAll("\\?", " ").replaceAll("\\*", " ");
 		query = query.toLowerCase().replaceAll("\\[", " ").replaceAll("\\]", " ");
@@ -51,5 +40,4 @@ public class QueryPreProcessor {
 		
 		return newQuery;
 	}
-
 }
