+++ pipeline/pom.xml
@@ -89,7 +89,11 @@
             <artifactId>illinois-md</artifactId>
             <version>3.1.32</version>
         </dependency>
-
+        <dependency>
+            <groupId>edu.illinois.cs.cogcomp</groupId>
+            <artifactId>illinois-datalessclassification</artifactId>
+            <version>3.1.32</version>
+        </dependency>
         <dependency>
             <groupId>edu.illinois.cs.cogcomp</groupId>
             <artifactId>illinois-srl</artifactId>
