+++ pipeline/src/main/java/edu/illinois/cs/cogcomp/pipeline/common/PipelineConfigurator.java
@@ -46,6 +46,8 @@ public class PipelineConfigurator extends AnnotatorServiceConfigurator {
     public static final Property USE_SRL_INTERNAL_PREPROCESSOR = new Property(
             SrlConfigurator.INSTANTIATE_PREPROCESSOR.key, FALSE);
     public static final Property USE_TIMEX3 = new Property("useTimex3", FALSE);
+    public static final Property USE_ESA_DATALESS = new Property("useESADataless", FALSE);
+    public static final Property USE_W2V_DATALESS = new Property("useW2VDataless", FALSE);
 
     /**
      * if 'true', the PipelineFactory will return a sentence-level pipeline that will use all viable
@@ -77,7 +79,8 @@ public class PipelineConfigurator extends AnnotatorServiceConfigurator {
                         USE_STANFORD_PARSE, USE_STANFORD_DEP, USE_SRL_VERB, USE_SRL_NOM, USE_SRL_PREP, USE_SRL_COMMA,
                         USE_QUANTIFIER, USE_VERB_SENSE, USE_JSON,
                         USE_LAZY_INITIALIZATION, USE_SRL_INTERNAL_PREPROCESSOR, SPLIT_ON_DASH,
-                        USE_SENTENCE_PIPELINE, USE_TIMEX3, USE_MENTION};
+                        USE_SENTENCE_PIPELINE, USE_TIMEX3, USE_MENTION,
+                        USE_ESA_DATALESS, USE_W2V_DATALESS};
         return (new AnnotatorServiceConfigurator().getConfig(new ResourceManager(
                 generateProperties(properties))));
     }
